Migrating from urllib2 to requests
Here is a third party module requests where saves from coding for a version, requests keeps it the same way regardless of the version of python. Is super easy and might save you some lines of codes.
http://docs.python-requests.org/en/master/user/quickstart/#make-a-request
This patch isnt working but gives you a reason to make it simple to create the get request. But self.opener.open().read() would need to change to